crawler-worker/requirements.txt
playwright>=1.40.0
redis>=5.0.0
python-dotenv>=1.0.0
requests>=2.31.0
playwright-stealth>=0.1.2
